# Make the stream parser iterative instead of recursive.