This makes the stream parsing code a lot less memory intensive, as the stack will be no longer deeply nested when parsing incoming data. It's not providing an obvious benefit in all cases, but it lays some groundwork that I'll be using for further changes.

This is the last thing I want to get in before I cut a `1.13.0` release.

Benchmarks:

Before:

```
Many result rows x 3.95 ops/sec ±1.91% (24 runs sampled)
Memory: 68.53125 MiB
inserting nvarchar(max) with 5242880 chars x 20.32 ops/sec ±2.63% (54 runs sampled)
Memory: 117.2109375 MiB
inserting nvarchar(max) with 4 chars x 534 ops/sec ±1.36% (82 runs sampled)
Memory: 22.62890625 MiB
inserting varbinary(4) with 4 bytes x 535 ops/sec ±1.37% (79 runs sampled)
Memory: 22.265625 MiB
inserting varbinary(max) with 50 MiB x 4.47 ops/sec ±1.78% (26 runs sampled)
Memory: 116.74609375 MiB
inserting varbinary(max) with 5 MiB x 44.24 ops/sec ±1.84% (73 runs sampled)
Memory: 86.34765625 MiB
inserting varbinary(max) with 4 bytes x 533 ops/sec ±1.37% (79 runs sampled)
Memory: 22.359375 MiB
parsing `COLMETADATA` tokens x 158 ops/sec ±2.45% (79 runs sampled)
Memory: 44.921875 MiB
parsing `DONEPROC` tokens x 145 ops/sec ±2.19% (79 runs sampled)
Memory: 43.9921875 MiB
parsing tokens for 100 rows x 178 ops/sec ±2.29% (75 runs sampled)
Memory: 46.34375 MiB
```

After:

```
Many result rows x 4.14 ops/sec ±1.38% (25 runs sampled)
Memory: 57.1796875 MiB
inserting nvarchar(max) with 5242880 chars x 19.05 ops/sec ±1.57% (50 runs sampled)
Memory: 75.10546875 MiB
inserting nvarchar(max) with 4 chars x 535 ops/sec ±1.40% (82 runs sampled)
Memory: 21.59375 MiB
inserting varbinary(4) with 4 bytes x 536 ops/sec ±1.19% (83 runs sampled)
Memory: 21.140625 MiB
inserting varbinary(max) with 50 MiB x 4.71 ops/sec ±2.15% (28 runs sampled)
Memory: 119.578125 MiB
inserting varbinary(max) with 5 MiB x 41.53 ops/sec ±1.49% (68 runs sampled)
Memory: 51.39453125 MiB
inserting varbinary(max) with 4 bytes x 533 ops/sec ±1.38% (85 runs sampled)
Memory: 20.91796875 MiB
parsing `COLMETADATA` tokens x 159 ops/sec ±2.72% (78 runs sampled)
Memory: 44.56640625 MiB
parsing `DONEPROC` tokens x 154 ops/sec ±2.12% (78 runs sampled)
Memory: 42.2578125 MiB
parsing tokens for 100 rows x 178 ops/sec ±1.89% (81 runs sampled)
Memory: 45.8671875 MiB
```